Changed FormDisplayDogs to stateless
conditions are not changeable on screen so no need for state
