+++ src/Components/FormDisplayDogs.js
@@ -1,57 +1,36 @@
 import React from 'react';
 
 import Dogs from '../Dogs';
-import toggleFilter from '../Filter.js';
 
-export default class FormDisplayDogs extends React.Component {
-	constructor(props) {
-		super(props);
+const FormDisplayDogs = props => {
+	const { username, house, kids, cats, dogs } = props;
+	const conditions = {};
+	// build conditions based on form input
+	// props.house can have a value of "house" or "apartment" or null
+	// props.kids can have a value of "kids" or "no kids" or null
 
-		this.state = {
-			conditions: {}
-		};
-		this.buildConditions(this.props);
+	if (kids === 'kids') {
+		// set good with kids condition
+		conditions.rescue_me_meta_good_with_kids = 'true';
+	}
+	if (cats === 'cats') {
+		// set good with kids condition
+		conditions.rescue_me_meta_good_with_cats = 'true';
 	}
 
-	buildConditions = props => {
-		// build conditions based on form input
-		// props.house can have a value of "house" or "apartment" or null
-		// props.kids can have a value of "kids" or "no kids" or null
+	if (house === 'apartment') {
+		//set small condition
+		conditions.rescue_me_meta_size = 'Small';
+	}
+	console.log(conditions);
 
-		if (props.kids === 'kids') {
-			// set good with kids condition
-			let group = 'rescue_me_meta_good_with_kids';
-			// this.setState({ conditions: group['true'] });
-			this.state.conditions[group] = 'true';
-			// toggle needs (conditions, setConditions, group, value)
-			// toggleFilter(this.state.conditions, this.setConditions, group, 'true');
-		}
-		if (props.cats === 'cats') {
-			// set good with kids condition
-			let group = 'rescue_me_meta_good_with_cats';
-			// this.setState({ conditions: group['true'] });
-			this.state.conditions[group] = 'true';
-		}
-		let group = 'rescue_me_meta_size';
-		if (props.house === 'apartment') {
-			//set small condition
-			// this.setState({ conditions: group['Small'] });
-			this.state.conditions[group] = 'Small';
-		}
-		console.log(this.state.conditions);
-	};
-
-	render() {
+	return (
 		// return <FilteredDogsList conditions={this.state.conditions} />;
-		return (
-			<div>
-				<Dogs
-					username={this.props.username}
-					dogs={this.props.dogs}
-					conditions={this.state.conditions}
-					setConditions={this.setConditions}
-				/>
-			</div>
-		);
-	}
-}
+
+		<div>
+			<Dogs username={username} dogs={dogs} conditions={conditions} />
+		</div>
+	);
+};
+
+export default FormDisplayDogs;
